Updated Howler to enable sound position (#1243)
Updated Howler to version 2.0.15

For what we use, we could use the "stereo()" function but as we already generated the coordinates, I settled for the "pos()" function.

My knowledge of sound isn't enough to understand where the "w" variable of the position goes (as "pos()" uses only x, y and z). My guess is that "pannerAttr()" might have the access to what we are looking for, but again I don't fully understand sound.